#71 cleaned files
Former-commit-id: 16e60d070c222c590f4c059946ff9d6ef2db2e50
