--- ReportGenerator/report_templates/html_report/carousel_sec_template.html
@@ -1,36 +1,0 @@
-<div class="row">
-    <div class="col col-md-12" >
-        <div class="row">
-            <div class="col col-md-8 offset-md-2">
-                <p><h3>SECTION_NAME</h3></p>      
-            </div>
-        </div>
-        <div class="row">
-            <div class="col col-md-2 offset-md-2">
-                <a href="SECTION_LINK_LINK">SECTION_LINK_TEXT</a>
-            </div>
-        </div>
-    </div>
-    <div class="col col-md-12">
-        <div class="row">
-        <div class="col col-md-8 offset-md-2">
-            <div id="SECTION_CAROUSEL_ID" class="carousel slide" data-ride="carousel" data-interval="false">
-                <ol class="carousel-indicators" style="background-color: black;opacity: 0.20;">
-SECTION_CAROUSEL_INDICATORS
-                </ol>
-                <div class="carousel-inner">
-SECTION_CAROUSEL_IMGS
-                </div>
-                <a class="carousel-control-prev" href="#SECTION_CAROUSEL_ID" role="button" data-slide="prev">
-                    <span class="carousel-control-prev-icon" aria-hidden="true" style="background-color: black;opacity: 0.35;"></span>
-                    <span class="sr-only">Previous</span>
-                </a>
-                <a class="carousel-control-next" href="#SECTION_CAROUSEL_ID" role="button" data-slide="next">
-                    <span class="carousel-control-next-icon" aria-hidden="true" style="background-color: black;opacity: 0.35;"></span>
-                    <span class="sr-only">Next</span>
-                </a>
-            </div>
-        </div>
-        </div>
-    </div>
-</div>
