--- ReportGenerator/report_templates/html_report/figure_template.html
@@ -1,8 +1,0 @@
-        <div class="row" >
-            <div class="row" >
-                <div class="col col-md-8 offset-md-2">
-                    <h4>FIGURE_NAME_HERE</h4>
-                </div>
-            </div>
-            <img class="img-fluid" src="FIGURE_PATH" alt="FIGURE_ALT_TEXT">
-        </div>
