--- ReportGenerator/report_templates/html_report/section_template.html
@@ -1,13 +1,0 @@
-    <div class="container">
-        <div class="row" >
-            <div class="col col-md-8 offset-md-2">
-                <p><h3>SECTION_NAME</h3></p>      
-            </div>
-            <div class="col col-md-2">
-                <a href="SECTION_LINK_LINK">SECTION_LINK_TEXT</a>
-            </div>
-        </div>
-        <div class="sub_sections">
-            SUB_SECTIONS_CONTENT
-        </div>
-    </div>
