--- ReportGenerator/report_templates/latex_report/buildDoc.bat
@@ -1,3 +1,0 @@
-pdflatex "acceptance_report.tex" > ar_first_Log.txt
-pdflatex "acceptance_report.tex" > ar_first_toc_Log.txt
-pdflatex "acceptance_report.tex" > ar_last_Log.txt
