+++ tests/test_data/cases/case_02_compound/Data/cross_section_locations.xyz
@@ -1,2 +1,34 @@
 name,x,y,length,branch,offset
-channel1_0.000, -24.0000, 75.0000, 3050.00, Channel1, 0.00
+channel1_0.000, -24.0000, 75.0000, 12.50, Channel1, 0.00
+channel1_25.000, 0.0000, 75.0000, 62.50, Channel1, 25.00
+channel1_125.000, 100.0000, 75.0000, 100.00, Channel1, 125.00
+channel1_225.000, 200.0000, 75.0000, 100.00, Channel1, 225.00
+channel1_325.000, 300.0000, 75.0000, 100.00, Channel1, 325.00
+channel1_425.000, 400.0000, 75.0000, 100.00, Channel1, 425.00
+channel1_525.000, 500.0000, 75.0000, 100.00, Channel1, 525.00
+channel1_625.000, 600.0000, 75.0000, 100.00, Channel1, 625.00
+channel1_725.000, 700.0000, 75.0000, 100.00, Channel1, 725.00
+channel1_825.000, 800.0000, 75.0000, 100.00, Channel1, 825.00
+channel1_925.000, 900.0000, 75.0000, 100.00, Channel1, 925.00
+channel1_1025.000, 1000.0000, 75.0000, 100.00, Channel1, 1025.00
+channel1_1125.000, 1100.0000, 75.0000, 100.00, Channel1, 1125.00
+channel1_1225.000, 1200.0000, 75.0000, 100.00, Channel1, 1225.00
+channel1_1325.000, 1300.0000, 75.0000, 100.00, Channel1, 1325.00
+channel1_1425.000, 1400.0000, 75.0000, 100.00, Channel1, 1425.00
+channel1_1525.000, 1500.0000, 75.0000, 100.00, Channel1, 1525.00
+channel1_1625.000, 1600.0000, 75.0000, 100.00, Channel1, 1625.00
+channel1_1725.000, 1700.0000, 75.0000, 100.00, Channel1, 1725.00
+channel1_1825.000, 1800.0000, 75.0000, 100.00, Channel1, 1825.00
+channel1_1925.000, 1900.0000, 75.0000, 100.00, Channel1, 1925.00
+channel1_2025.000, 2000.0000, 75.0000, 100.00, Channel1, 2025.00
+channel1_2125.000, 2100.0000, 75.0000, 100.00, Channel1, 2125.00
+channel1_2225.000, 2200.0000, 75.0000, 100.00, Channel1, 2225.00
+channel1_2325.000, 2300.0000, 75.0000, 100.00, Channel1, 2325.00
+channel1_2425.000, 2400.0000, 75.0000, 100.00, Channel1, 2425.00
+channel1_2525.000, 2500.0000, 75.0000, 100.00, Channel1, 2525.00
+channel1_2625.000, 2600.0000, 75.0000, 100.00, Channel1, 2625.00
+channel1_2725.000, 2700.0000, 75.0000, 100.00, Channel1, 2725.00
+channel1_2825.000, 2800.0000, 75.0000, 100.00, Channel1, 2825.00
+channel1_2925.000, 2900.0000, 75.0000, 100.00, Channel1, 2925.00
+channel1_3025.000, 3000.0000, 75.0000, 62.50, Channel1, 3025.00
+channel1_3050.000, 3024.0000, 75.0000, 12.50, Channel1, 3050.00
